Move loadEnvironment.js out of /backend
