+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "dotenv": "^16.3.1"
+  }
+}
